Move all components `desktop` package.
