+++ desktop/src/main/kotlin/io/spine/examples/pingh/desktop/Icons.kt
@@ -24,29 +24,27 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.pingh.desktop.component
+package io.spine.examples.pingh.desktop
 
-import androidx.compose.foundation.shape.CircleShape
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.clip
-import coil3.compose.AsyncImage
-import io.spine.net.Url
+import androidx.compose.ui.graphics.painter.BitmapPainter
+import androidx.compose.ui.res.loadImageBitmap
+import androidx.compose.ui.res.useResource
 
 /**
- * Displays a round avatar with an image loaded from the specified URL.
+ * Icons used in the Pingh desktop client.
  *
- * The image loads asynchronously.
+ * These icons are [free](https://bit.ly/3Tjmsqg) for personal and commercial
+ * use with attribution.
+ *
+ * Created by [Icon Hubs](https://www.flaticon.com/authors/icon-hubs).
  */
-@Composable
-public fun Avatar(
-    url: Url,
-    modifier: Modifier = Modifier
-) {
-    AsyncImage(
-        model = url.spec,
-        contentDescription = null,
-        modifier = modifier
-            .clip(CircleShape)
-    )
+public object Icons {
+    public val snooze: BitmapPainter =
+        BitmapPainter(useResource("icons/snooze.png", ::loadImageBitmap))
+    public val profile: BitmapPainter =
+        BitmapPainter(useResource("icons/profile.png", ::loadImageBitmap))
+    public val refresh: BitmapPainter =
+        BitmapPainter(useResource("icons/refresh.png", ::loadImageBitmap))
+    public val back: BitmapPainter =
+        BitmapPainter(useResource("icons/back.png", ::loadImageBitmap))
 }
