+++ desktop/src/main/kotlin/io/spine/examples/pingh/desktop/Mentions.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.examples.pingh.desktop.home
+package io.spine.examples.pingh.desktop
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -46,6 +46,8 @@ import androidx.compose.material3.ElevatedCard
 import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
@@ -54,14 +56,12 @@ import androidx.compose.ui.geometry.Offset
 import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.platform.LocalUriHandler
 import androidx.compose.ui.unit.dp
+import com.google.protobuf.util.Timestamps
 import io.spine.examples.pingh.client.DesktopClient
-import io.spine.examples.pingh.desktop.component.Avatar
-import io.spine.examples.pingh.desktop.component.IconButton
-import io.spine.examples.pingh.desktop.component.Icons
-import io.spine.examples.pingh.desktop.toDatetime
-import io.spine.examples.pingh.desktop.truncate
+import io.spine.examples.pingh.mentions.MentionId
 import io.spine.examples.pingh.mentions.MentionStatus
 import io.spine.examples.pingh.mentions.MentionView
+import java.lang.Thread.sleep
 
 /**
  * Maximum length for the title of the mention card.
@@ -71,6 +71,15 @@ import io.spine.examples.pingh.mentions.MentionView
 private const val maxLengthOfMentionCardTitle = 27
 
 /**
+ * Order of statuses for sorting mentions in the [sortByStatusAndWhenMentioned] method.
+ */
+private val statusOrder = mapOf(
+    MentionStatus.UNREAD to 0,
+    MentionStatus.SNOOZED to 1,
+    MentionStatus.READ to 2
+)
+
+/**
  * Displays the 'Home' page in the application.
  *
  * This page is the main interface where users can manage their mentions.
@@ -78,7 +87,7 @@ private const val maxLengthOfMentionCardTitle = 27
  * possible to manually update the list of mentions from the server.
  */
 @Composable
-public fun HomePage(client: DesktopClient) {
+internal fun HomePage(client: DesktopClient) {
     val model = remember { HomePageModel(client) }
     Column(
         Modifier.fillMaxSize()
@@ -264,3 +273,98 @@ private fun SnoozeButton(model: HomePageModel, mention: MentionView) {
         else -> {}
     }
 }
+
+/**
+ * UI Model for the [HomePage].
+ *
+ * UI Model is a layer between `@Composable` functions and client.
+ */
+private class HomePageModel(private val client: DesktopClient) {
+
+    private companion object {
+        /**
+         * Delay before reading mentions so that the read-side on the server can be updated.
+         *
+         * Time is specified in milliseconds.
+         */
+        private const val delayBeforeReadingMentions = 100L
+    }
+
+    /**
+     * User mentions.
+     */
+    private var mentions: MutableState<MentionsList> = mutableStateOf(client.findUserMentions())
+
+    /**
+     * Returns all user mentions.
+     */
+    public fun mentions(): MentionsList = mentions.value
+
+    /**
+     * Updates the user mentions.
+     */
+    public fun updateMentions() {
+        client.updateMentions(
+            onSuccess = {
+                sleep(delayBeforeReadingMentions)
+                mentions.value = client.findUserMentions()
+            }
+        )
+    }
+
+    /**
+     * Marks the mention as snoozed.
+     */
+    public fun markMentionAsSnoozed(id: MentionId) {
+        client.markMentionAsSnoozed(id) {
+            mentions.value = mentions.value.setMentionStatus(id, MentionStatus.SNOOZED)
+        }
+    }
+
+    /**
+     * Marks that the mention is read.
+     */
+    public fun markMentionAsRead(id: MentionId) {
+        client.markMentionAsRead(id) {
+            mentions.value = mentions.value.setMentionStatus(id, MentionStatus.READ)
+        }
+    }
+}
+
+/**
+ * List of `MentionsView`s.
+ */
+private typealias MentionsList = List<MentionView>
+
+/**
+ * Creates a new list by replacing the status of one mention with another.
+ */
+private fun MentionsList.setMentionStatus(id: MentionId, status: MentionStatus): MentionsList {
+    val idInList = this.indexOfFirst { it.id == id }
+    val updatedMention = this[idInList]
+        .toBuilder()
+        .setStatus(status)
+        .vBuild()
+    val newMentionsList = this.toMutableList()
+    newMentionsList[idInList] = updatedMention
+    return newMentionsList
+}
+
+/**
+ * Returns a `MentionsList` sorted such that unread mentions come first,
+ * followed by snoozed mentions, and read mentions last.
+ *
+ * Within each group, mentions are sorted by the time they were made.
+ *
+ * @see [statusOrder]
+ */
+private fun MentionsList.sortByStatusAndWhenMentioned(): MentionsList =
+    this.sortedWith { firstMentions, secondMentions ->
+        val statusComparison = statusOrder[firstMentions.status]!!
+            .compareTo(statusOrder[secondMentions.status]!!)
+        if (statusComparison != 0) {
+            statusComparison
+        } else {
+            Timestamps.compare(secondMentions.whenMentioned, firstMentions.whenMentioned)
+        }
+    }
